requirements.txt
Django~=4.2
django-embed-video~=1.4.8
requests~=2.28.2
tensorflow~=2.12.0
numpy~=1.23.5
keras~=2.12.0
Pillow~=9.5.0
pytube~=12.1.3
django-embed-video